+++ CHANGELOG.md
@@ -27,4 +27,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added interface `Format\TimeFormatter`.
 - Added interface `Format\TimeParser`.
 - Added exception `Exception\FormatException`.
-- Added `Format\SimpleDateFormatter` which makes formatting of a date to the format `YYYY-MM-DD`.
+- Added `Format\SimpleDateFormatter` which makes formatting of date to the format `YYYY-MM-DD`.
+- Added `Format\SimpleTimeFormatter` which makes formatting of time to the format `HH:mm:ss`.